MapReduce works remotely for several slaves with the CC files
- changed <String, Integer> to <String, Long>, will maybe be useful
- stopped sending results over TCP as it didn't work
- the file to mapReduce in an entry parameter for the slave

+++ slave/src/messages/ReduceResult.java
@@ -4,15 +4,15 @@ import java.io.Serializable;
 import java.util.HashMap;
 
 public class ReduceResult extends Message implements Serializable {
-    private HashMap<String, Integer> wordCount;
+    private HashMap<String, Long> wordCount;
     private int id;
 
-    public ReduceResult(int id, HashMap<String, Integer> wordCount) {
+    public ReduceResult(int id, HashMap<String, Long> wordCount) {
         this.id = id;
         this.wordCount = wordCount;
     }
 
-    public HashMap<String, Integer> getWordCount() {
+    public HashMap<String, Long> getWordCount() {
         return wordCount;
     }
 }
